backend/requirements.txt

Flask==2.3.3
Flask-CORS==3.0.10
Flask-SQLAlchemy==3.0.5
email-validator==2.0.0.post2
SQLAlchemy==1.4.54
python-dotenv==1.0.0